+++ examples/v0.71.19/babel.config.js
@@ -7,7 +7,7 @@ module.exports = {
       {
         envName: 'APP_ENV',
         moduleName: '@env',
-        allowlist: ['HOT_UPDATER_SUPABASE_URL'],
+        allowlist: ['HOT_UPDATER_SUPABASE_URL', 'HOT_UPDATER_SENTRY_DSN'],
         path: '.env',
       },
     ],