+++ packages/react-native/android/src/oldarch/HotUpdaterSpec.kt
@@ -18,4 +18,9 @@ abstract class HotUpdaterSpec internal constructor(
     abstract fun reload()
 
     abstract fun getAppVersion(promise: Promise)
+
+    abstract fun setChannel(
+        channel: String,
+        promise: Promise,
+    )
 }