+++ packages/react-native/src/index.ts
@@ -7,6 +7,7 @@ import {
   getMinBundleId,
   notifyAppReady,
   reload,
+  setChannel,
   updateBundle,
 } from "./native";
 import { runUpdateProcess } from "./runUpdateProcess";
@@ -68,8 +69,9 @@ export const HotUpdater = {
    * Fetches the current release channel of the app.
    *
    * By default, if no channel is specified, the app is assigned to the 'production' channel.
+   * If the app is running in development mode, the channel will be `null`.
    *
-   * @returns {string} The current release channel of the app
+   * @returns {string | null} The current release channel of the app
    *
    * @example
    * ```ts
@@ -78,6 +80,10 @@ export const HotUpdater = {
    * ```
    */
   getChannel,
+  /**
+   * Sets the channel for the app.
+   */
+  setChannel,
   /**
    * Adds a listener to HotUpdater events.
    *